app/requirements.txt
# pip install -r requirements.txt
requests
python-dotenv